+++ src/prefabs/emailinput.tsx
@@ -1,0 +1,42 @@
+import * as React from 'react';
+import { BeforeCreateArgs, Icon, prefab } from '@betty-blocks/component-sdk';
+import { TextInput } from './structures/TextInput';
+
+const beforeCreate = ({
+  close,
+  components: { CreateFormInputWizard },
+  prefab,
+  save,
+}: BeforeCreateArgs) => {
+  const actionVariableOption = prefab.structure[0].options.find(
+    (option: { type: string }) => option.type === 'ACTION_JS_VARIABLE',
+  );
+
+  if (!actionVariableOption) {
+    return <div>Prefab is missing the actionVariable component option</div>;
+  }
+
+  return (
+    <CreateFormInputWizard
+      supportedKinds={['EMAIL_ADDRESS']}
+      actionVariableOption={actionVariableOption.key}
+      labelOptionKey="label"
+      nameOptionKey="actionVariableId"
+      close={close}
+      prefab={prefab}
+      save={save}
+    />
+  );
+};
+
+const attributes = {
+  category: 'FORMV2',
+  icon: Icon.EmailInputIcon,
+  keywords: ['Form', 'input'],
+};
+
+const pattern = '[a-z0-9._%+-]+@[a-z0-9.-]+[\\.][a-z]{2,4}$';
+
+export default prefab('Email v2', attributes, beforeCreate, [
+  TextInput({ label: 'email', type: 'email', pattern, adornmentIcon: 'Email' }),
+]);