+++ src/prefabs/pageWithCardAndListView.js
@@ -1189,7 +1189,7 @@
                         type: 'SIZE',
                         label: 'Width',
                         key: 'width',
-                        value: '',
+                        value: '100%',
                         configuration: {
                           as: 'UNIT',
                         },
@@ -2826,7 +2826,7 @@
                         type: 'SIZE',
                         label: 'Width',
                         key: 'width',
-                        value: '',
+                        value: '100%',
                         configuration: {
                           as: 'UNIT',
                         },
@@ -14782,7 +14782,7 @@
                         type: 'SIZE',
                         label: 'Width',
                         key: 'width',
-                        value: '',
+                        value: '100%',
                         configuration: {
                           as: 'UNIT',
                         },