+++ src/prefabs/pageWithEditDatatable.js
@@ -1003,7 +1003,7 @@
                     type: 'SIZE',
                     label: 'Width',
                     key: 'width',
-                    value: '',
+                    value: '100%',
                     configuration: {
                       as: 'UNIT',
                     },
@@ -2640,7 +2640,7 @@
                     type: 'SIZE',
                     label: 'Width',
                     key: 'width',
-                    value: '',
+                    value: '100%',
                     configuration: {
                       as: 'UNIT',
                     },
@@ -7236,7 +7236,7 @@
                     type: 'SIZE',
                     label: 'Width',
                     key: 'width',
-                    value: '',
+                    value: '100%',
                     configuration: {
                       as: 'UNIT',
                     },